Move some debug/input/outp functions
