+++ CrusherChess/src/CrusherChess.cpp
@@ -24,189 +24,9 @@
 #include "uci.h"
 
 
-class TextAttr
-{
-	friend std::ostream& operator<<(std::ostream& out, TextAttr attr)
-	{
-		SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), attr.value);
-		return out;
-	}
-public:
-	explicit TextAttr(WORD attributes) : value(attributes) {}
-private:
-	WORD value;
-};
-
-#define FOREGROUND_WHITE (FOREGROUND_RED | FOREGROUND_GREEN | FOREGROUND_BLUE)
-#define BACKGROUND_WHITE (BACKGROUND_RED | BACKGROUND_GREEN | BACKGROUND_BLUE)
-
 
 #pragma warning(disable:6031)
 
-
-
-/****************************************\
- ========================================
-		   Debug + Input & output
- ========================================
-\****************************************/
-#pragma region
-
-int get_piece(int file, int rank)
-{
-	int piece = -1;
-
-	for (int i = P; i <= k; i++)
-	{
-		if (get_bit(bitboards[i], (rank * 8 + file)))
-		{
-			piece = i;
-			break;
-		}
-	}
-	return piece;
-}
-
-std::string get_fen()
-{
-	std::ostringstream ss;
-
-	for (int r = 0; r < 8; r++)
-	{
-		for (int f = 0; f < 8; f++)
-		{
-			int empty = 0;
-
-			while (get_piece(f, r) == -1 && f++ < 8)
-				empty++;
-
-			if (empty)
-				ss << empty;
-
-			if (f <= 7)
-			{
-				int piece = get_piece(f, r);
-
-				if (piece == -1)
-					assert(!"piece is -1");
-
-				ss << ascii_pieces[piece];
-			}
-		}
-
-		if (r < 7)
-			ss << '/';
-	}
-
-	ss << (side == WHITE ? " w " : " b ");
-
-	if (castle & 1)		ss << 'K';
-	if (castle & 2)		ss << 'Q';
-	if (castle & 4)		ss << 'k';
-	if (castle & 8)		ss << 'q';
-	if (!(castle & 15))	ss << '-';
-
-	ss << (enpassant == NO_SQ ? " - " : " " + std::string(square_to_coordinates[enpassant]) + " ");
-
-	//need to add half and full move counters
-	ss << "0 " << "1";
-
-	return ss.str();
-}
-
-void print_bitboard(U64 bitboard)
-{
-	wprintf(L"    Bitboard: %llu\n", bitboard);
-	for (int rank = 0, i = 0; rank < 8; ++rank)
-	{
-		wprintf(L" %d  ", 8 - rank);
-
-		for (int file = 0; file < 8; ++file, ++i)
-			wprintf(L"%llu ", get_bit(bitboard, i));
-
-		wprintf(L"\n");
-	}
-	//print board files
-	wprintf(L"    a b c d e f g h\n\n");
-	//print bitboard as unsigned number  
-}
-
-void print_board(bool ascii = false)
-{
-	printf("\n");
-	HANDLE handle = GetStdHandle(STD_OUTPUT_HANDLE);
-	DWORD written = 0;
-
-	for (int rank = 0, square = 0; rank < 8; rank++)
-	{
-		std::cout << TextAttr(FOREGROUND_WHITE);
-		printf("%d ", 8 - rank);
-
-		for (int file = 0; file < 8; file++, square++)
-		{
-			int piece = -1;
-
-			//loop over all the pieces
-			for (int bb_piece = P; bb_piece <= k; bb_piece++)
-			{
-				if (get_bit(bitboards[bb_piece], square))
-					piece = bb_piece;
-			}
-
-			((square + rank + 1) % 2) ? std::cout << TextAttr(BACKGROUND_WHITE) : std::cout << TextAttr(FOREGROUND_WHITE);
-
-			if (piece != -1)
-			{
-				if (ascii)
-					printf("%c ", ascii_pieces[piece]);
-				else
-				{
-					if ((square + rank + 1) % 2)
-					{
-						if (piece > 5)
-							piece -= 6;
-						else
-							piece += 6;
-					}
-
-					//WriteConsoleW(handle, unicode_pieces[piece], 1, &written, NULL);
-
-//					if (piece == P)
-//						WriteConsoleW(handle, unicode_pieces[piece], 1, &written, NULL);
-//					else
-					{
-						WriteConsoleW(handle, unicode_pieces[piece], 1, &written, NULL);
-						WriteConsoleW(handle, L" ", 1, &written, NULL);
-					}
-				}
-			}
-			else
-				printf("  ");
-		}
-		std::cout << TextAttr(FOREGROUND_WHITE);
-		printf("\n");
-	}
-	printf("  a b c d e f g h\n\n");
-
-	printf("Side: %s\n", side ? "Black" : "White");
-
-	printf("En Passant: %s\n", (enpassant != NO_SQ) ? (square_to_coordinates[enpassant]) : "no");
-
-	printf("Castling: %c%c%c%c\n",
-		(castle & WK) ? 'K' : '-',
-		(castle & WQ) ? 'Q' : '-',
-		(castle & BK) ? 'k' : '-',
-		(castle & BQ) ? 'q' : '-'
-	);
-
-	std::cout << "Fen: " << get_fen() << custom_endl;
-
-	printf("Key: %llx\n\n", hash_key);
-}
-
-
-
-#pragma endregion
 
 /****************************************\
  ========================================
@@ -300,8 +120,8 @@ void init_all()
 	//init hash table with default 64mb
 	init_hash_table(64);
 
-    throw new std::runtime_error("Not implemented");
-	//parse_fen(start_position);
+   
+	parse_fen(start_position);
 }
 #pragma endregion
